requirements.txt
torch>=1.13.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
pillow
numpy==1.26.0
pandas
tqdm
scikit-learn
gradio>=3.20
protobuf
opencv-python-headless
flask